修正 ディレクトリ構成
